+++ src/mediarithmics/plugins/email-renderer/template/EmailRendererTemplatePlugin.ts
@@ -13,17 +13,6 @@ export interface EmailRendererTemplateInstanceContext
 }
 
 export abstract class EmailRendererTemplatePlugin extends EmailRendererPlugin<EmailRendererTemplateInstanceContext> {
-  /**
- * Helper to fetch the content of a template
- * @param templatePath  The raw (e.g. non URL encoded) mics URI to the template file as a string.
- * @returns       A Buffer with the file content in it. This have to be decoded with the proper encoding.
- */
-  async fetchTemplateContent(templatePath: string): Promise<Buffer> {
-    const templateContent = await super.fetchDataFile(templatePath);
-
-    this.logger.debug(`Fetched template : ${templateContent}`);
-    return templateContent;
-  }
 
   /**
    * The engineBuilder that can be used to compile the template
@@ -31,7 +20,7 @@ export abstract class EmailRendererTemplatePlugin extends EmailRendererPlugin<Em
    * 
    * Have to be overriden (see examples)
    */
-  protected engineBuilder: ExploreableInternalsTemplatingEngine<any, any, any, any> & ProfileDataTemplater;
+  protected abstract engineBuilder: ExploreableInternalsTemplatingEngine<any, any, any, any> & ProfileDataTemplater;
 
   constructor(enableThrottling = false) {
     super(enableThrottling);