+++ src/tests/AdRendererBaseTests.ts
@@ -90,6 +90,12 @@ describe("Fetch DisplayAd API", () => {
       done();
     });
   });
+
+  afterEach(() => {
+    // We clear the cache so that we don't have any processing still running in the background
+    runner.plugin.pluginCache.clear();
+  });
+
 });
 
 describe("Ad Contents API test", function() {
@@ -109,7 +115,8 @@ describe("Ad Contents API test", function() {
   }
 
   const plugin = new MyFakeAdRenderer2(false);
-
+  let runner: core.TestingPluginRunner;
+  
   it("Check that the plugin is giving good results with a simple adContents handler", function(
     done
   ) {
@@ -164,7 +171,7 @@ describe("Ad Contents API test", function() {
       })
     );
 
-    const runner = new core.TestingPluginRunner(plugin, rpMockup);
+    runner = new core.TestingPluginRunner(plugin, rpMockup);
 
     request(runner.plugin.app)
       .post("/v1/init")
@@ -202,8 +209,13 @@ describe("Ad Contents API test", function() {
         
       });
 
-
   });
+
+  afterEach(() => {
+    // We clear the cache so that we don't have any processing still running in the background
+    runner.plugin.pluginCache.clear();
+  });
+  
 });
 
 describe("Instance Context check", () => {